Added .gitignore and project configuration files
This update includes the addition of several new files to the project. These include a .gitignore file, which specifies files and directories that should be ignored by Git, as well as several configuration files for the IDE.

In addition, changes were made to main.go. A time tracking feature was added to the counter struct along with an SN field. The increaseCounter function now also calculates time difference between calls and accepts an SN query parameter. Error handling was also improved in the main function where it now logs fatal errors when starting up the server.